feat(minijinja): deprecate contrib module on v2
