+++ docs/reference/plugins/minijinja.rst
@@ -1,0 +1,6 @@
+=========
+minijinja
+=========
+
+.. automodule:: litestar.plugins.minijinja
+    :members:
